+++ tests/tests/wgpu-gpu/device.rs
@@ -511,12 +511,14 @@ static DEVICE_DESTROY_THEN_MORE: GpuTestConfiguration = GpuTestConfiguration::ne
         fail(
             &ctx.device,
             || unsafe {
-                let _ = ctx
-                    .device
-                    .create_shader_module_spirv(&wgpu::ShaderModuleDescriptorSpirV {
-                        label: None,
-                        source: std::borrow::Cow::Borrowed(&[]),
-                    });
+                let _ = ctx.device.create_shader_module_passthrough(
+                    wgpu::ShaderModuleDescriptorPassthrough::SpirV(
+                        wgpu::ShaderModuleDescriptorSpirV {
+                            label: None,
+                            source: std::borrow::Cow::Borrowed(&[]),
+                        },
+                    ),
+                );
             },
             Some("device with '' label is invalid"),
         );